Replace Discord link w/ additional context
We use the TAGNW Discord for hosting conversations related to the OWASP
Bellingham Chapter, and I don't see anything in the chapter guidelines against
this type of co-existence.

Added additional context to the link so that it's clear which channel the
OWASP Bellingham discussions are held in. Hopefully this clarifies that it is
not simply a third party link, and is indeed directly related to the chapter.

Also, fixing commit history via force push as Github appears to have committed
my update twice by accident.